+++ pyproject.toml
@@ -100,7 +100,9 @@ unit-test = [
   "syrupy>=5.1,<6",
   "polyfactory>=3.2,<4",
   "xmltodict-rs>=0.13,<0.14",
-  "orjson>=3.11,<4",
+  # We don't install `orjson` on 3.14, because we run `3.14t` as well,
+  # but no way to disable it specifically for `t` version :(
+  "orjson>=3.11,<4 ; python_version < '3.14'",
 ]
 integration = [
   "tracecov>=0.19,<0.20",
